refactor: Docs 폴더 이동
