+++ src/main/java/oncog/cogroom/domain/auth/controller/docs/AuthControllerDocs.java
@@ -1,4 +1,4 @@
-package oncog.cogroom.domain.auth.docs;
+package oncog.cogroom.domain.auth.controller.docs;
 
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
