+++ src/main/java/oncog/cogroom/domain/member/controller/MemberController.java
@@ -3,10 +3,8 @@ package oncog.cogroom.domain.member.controller;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import oncog.cogroom.domain.member.docs.MemberControllerDocs;
+import oncog.cogroom.domain.member.controller.docs.MemberControllerDocs;
 import oncog.cogroom.domain.member.dto.MemberRequestDTO;
-import oncog.cogroom.domain.member.dto.MemberResponseDTO;
-import oncog.cogroom.domain.member.entity.Member;
 import oncog.cogroom.domain.member.service.MemberService;
 import oncog.cogroom.global.common.response.ApiResponse;
 import oncog.cogroom.global.common.response.code.ApiSuccessCode;
